[ServerManage] Cog model restructure
Restructure cog `ServerManage` according to SFUAnime/Ren#608 by:
- Moving all methods from class `ServerManage` to `Core`
- Decoupling core command handling method calls from Red's command
  handler decorators by splitting `commands.py` into `commandsCore.py`
  and `commandHandlers.py`
- Dropping `ServerManageMeta` metaclass

No function changes are made in this commit.
