+++ servermanage/commandHandlers.py
@@ -1,28 +1,22 @@
-from abc import ABC
-from redbot.core import Config, checks, commands, data_manager
+from redbot.core import checks, commands
 from redbot.core.commands.context import Context
-from redbot.core.utils import AsyncIter
-from redbot.core.utils.chat_formatting import error, pagify, warning
-from redbot.core.utils.menus import DEFAULT_CONTROLS, menu
-from redbot.core.bot import Red
 
-from .constants import IMAGE_BANNERS
-from .meta import ServerManageMeta
+from .commandsCore import CommandsCore
 
 
-class ServerManageCommands(ABC, metaclass=ServerManageMeta):
+class CommandHandlers(CommandsCore):
     @commands.group(name="servermanage")
     @commands.guild_only()
     @checks.mod_or_permissions()
-    async def serverManage(self, ctx: Context):
+    async def _grpServerManage(self, ctx: Context):
         """Manage server icons and banners."""
 
-    @serverManage.group(name="icons", aliases=["icon"])
-    async def serverIcons(self, ctx: Context):
+    @_grpServerManage.group(name="icons", aliases=["icon"])
+    async def _grpServerIcons(self, ctx: Context):
         """Manage server icons."""
 
-    @serverIcons.command(name="add", aliases=["create"])
-    async def iconAdd(self, ctx: Context, iconName: str):
+    @_grpServerIcons.command(name="add", aliases=["create"])
+    async def _cmdIconAdd(self, ctx: Context, iconName: str):
         """Add a server icon to the database.
 
         Parameters
@@ -32,10 +26,10 @@ class ServerManageCommands(ABC, metaclass=ServerManageMeta):
         image: attachment
             The server icon, included as an attachment.
         """
-        return await self.imageAdd(ctx, iconName, imageType="icons")
+        return await self.cmdIconAdd(ctx=ctx, iconName=iconName)
 
-    @serverIcons.command(name="remove", aliases=["del", "delete", "rm"])
-    async def iconRemove(self, ctx: Context, iconName: str):
+    @_grpServerIcons.command(name="remove", aliases=["del", "delete", "rm"])
+    async def _cmdIconRemove(self, ctx: Context, iconName: str):
         """Remove a server icon from the database.
 
         Parameters
@@ -43,10 +37,10 @@ class ServerManageCommands(ABC, metaclass=ServerManageMeta):
         iconName: str
             The icon name you wish to remove.
         """
-        return await self.imageRemove(ctx, iconName)
+        return await self.cmdIconRemove(ctx=ctx, iconName=iconName)
 
-    @serverIcons.command(name="show")
-    async def iconShow(self, ctx: Context, iconName: str):
+    @_grpServerIcons.command(name="show")
+    async def _cmdIconShow(self, ctx: Context, iconName: str):
         """Show a server icon from the database.
 
         Parameters
@@ -54,15 +48,15 @@ class ServerManageCommands(ABC, metaclass=ServerManageMeta):
         iconName: str
             The icon name you wish to show.
         """
-        await self.imageShow(ctx, iconName)
+        await self.cmdIconShow(ctx=ctx, iconName=iconName)
 
-    @serverIcons.command(name="list", aliases=["ls"])
-    async def iconList(self, ctx: Context):
+    @_grpServerIcons.command(name="list", aliases=["ls"])
+    async def _cmdIconList(self, ctx: Context):
         """List the server icons associated with each date."""
-        return await self.imageList(ctx)
+        return await self.cmdIconList(ctx=ctx)
 
-    @serverIcons.command(name="set")
-    async def iconSet(self, ctx: Context, month: int, day: int, iconName: str):
+    @_grpServerIcons.command(name="set")
+    async def _cmdIconSet(self, ctx: Context, month: int, day: int, iconName: str):
         """Set when to change the server icon.
 
         Parameters
@@ -74,10 +68,10 @@ class ServerManageCommands(ABC, metaclass=ServerManageMeta):
         iconName: str
             The name of the server icon to change to. The icon should already be added.
         """
-        await self.imageDateSet(ctx, month, day, iconName)
+        await self.cmdIconSet(ctx=ctx, month=month, day=day, iconName=iconName)
 
-    @serverIcons.command(name="reset")
-    async def iconReset(self, ctx: Context, month: int, day: int):
+    @_grpServerIcons.command(name="reset")
+    async def _cmdIconReset(self, ctx: Context, month: int, day: int):
         """Remove a date when to change the server icon.
 
         Parameters
@@ -87,14 +81,14 @@ class ServerManageCommands(ABC, metaclass=ServerManageMeta):
         day: int
             The day of the month to remove any server icon changes, expressed as a number.
         """
-        await self.imageDateReset(ctx, month, day)
+        await self.cmdIconReset(ctx=ctx, month=month, day=day)
 
-    @serverManage.group(name="banners", aliases=["banner"])
-    async def serverBanners(self, ctx: Context):
+    @_grpServerManage.group(name="banners", aliases=["banner"])
+    async def _grpServerBanners(self, ctx: Context):
         """Manage server banners."""
 
-    @serverBanners.command(name="add", aliases=["create"])
-    async def bannerAdd(self, ctx: Context, bannerName: str):
+    @_grpServerBanners.command(name="add", aliases=["create"])
+    async def _cmdBannerAdd(self, ctx: Context, bannerName: str):
         """Add a server banner to the database.
 
         Parameters
@@ -104,10 +98,10 @@ class ServerManageCommands(ABC, metaclass=ServerManageMeta):
         image: attachment
             The server banner, included as an attachment.
         """
-        return await self.imageAdd(ctx, bannerName, imageType=IMAGE_BANNERS)
+        return await self.cmdBannerAdd(ctx=ctx, bannerName=bannerName)
 
-    @serverBanners.command(name="remove", aliases=["del", "delete", "rm"])
-    async def bannerRemove(self, ctx: Context, bannerName: str):
+    @_grpServerBanners.command(name="remove", aliases=["del", "delete", "rm"])
+    async def _cmdBannerRemove(self, ctx: Context, bannerName: str):
         """Remove a server banner from the database.
 
         Parameters
@@ -115,10 +109,10 @@ class ServerManageCommands(ABC, metaclass=ServerManageMeta):
         bannerName: str
             The banner name you wish to remove.
         """
-        return await self.imageRemove(ctx, bannerName, imageType=IMAGE_BANNERS)
+        return await self.cmdBannerRemove(ctx=ctx, bannerName=bannerName)
 
-    @serverBanners.command(name="show")
-    async def bannerShow(self, ctx: Context, bannerName: str):
+    @_grpServerBanners.command(name="show")
+    async def _cmdBannerShow(self, ctx: Context, bannerName: str):
         """Show a server banner from the database.
 
         Parameters
@@ -126,15 +120,15 @@ class ServerManageCommands(ABC, metaclass=ServerManageMeta):
         bannerName: str
             The banner name you wish to show.
         """
-        await self.imageShow(ctx, bannerName, imageType=IMAGE_BANNERS)
+        await self.cmdBannerShow(ctx=ctx, bannerName=bannerName)
 
-    @serverBanners.command(name="list", aliases=["ls"])
-    async def bannerList(self, ctx: Context):
+    @_grpServerBanners.command(name="list", aliases=["ls"])
+    async def _cmdBannerList(self, ctx: Context):
         """List the server banners associated with each date."""
-        return await self.imageList(ctx, imageType=IMAGE_BANNERS)
+        return await self.cmdBannerList(ctx=ctx)
 
-    @serverBanners.command(name="set")
-    async def bannerSet(self, ctx: Context, month: int, day: int, bannerName: str):
+    @_grpServerBanners.command(name="set")
+    async def _cmdBannerSet(self, ctx: Context, month: int, day: int, bannerName: str):
         """Set when to change the server banner.
 
         Parameters
@@ -146,10 +140,10 @@ class ServerManageCommands(ABC, metaclass=ServerManageMeta):
         bannerName: str
             The name of the server banner to change to. The banner should already be added.
         """
-        await self.imageDateSet(ctx, month, day, bannerName, imageType=IMAGE_BANNERS)
+        await self.cmdBannerSet(ctx=ctx, month=month, day=day, bannerName=bannerName)
 
-    @serverBanners.command(name="reset")
-    async def bannerReset(self, ctx: Context, month: int, day: int):
+    @_grpServerBanners.command(name="reset")
+    async def _cmdBannerReset(self, ctx: Context, month: int, day: int):
         """Remove a date when to change the server banner.
 
         Parameters
@@ -159,4 +153,4 @@ class ServerManageCommands(ABC, metaclass=ServerManageMeta):
         day: int
             The day of the month to remove any server banner changes, expressed as a number.
         """
-        await self.imageDateReset(ctx, month, day, imageType=IMAGE_BANNERS)
+        await self.cmdBannerReset(ctx=ctx, month=month, day=day)
