--- servermanage/meta.py
@@ -1,6 +1,0 @@
-from abc import ABC
-from redbot.core import commands
-
-
-class ServerManageMeta(type(commands.Cog), type(ABC)):
-    pass
